Move all Optimize components into a components subfolder.
