+++ assets/js/modules/optimize/components/common/AMPExperimentJSONField.js
@@ -29,14 +29,14 @@ import Data from 'googlesitekit-data';
 import {
 	Input,
 	TextField,
-} from '../../../material-components';
+} from '../../../../material-components';
 import classnames from 'classnames';
-import { STORE_NAME } from '../datastore/constants';
-import { STORE_NAME as CORE_SITE } from '../../../googlesitekit/datastore/site/constants';
-import { STORE_NAME as MODULES_ANALYTICS } from '../../analytics/datastore/constants';
-import { isValidAMPExperimentJSON } from '../util';
-import Link from '../../../components/link';
-import ErrorText from '../../../components/error-text';
+import { STORE_NAME } from '../../datastore/constants';
+import { STORE_NAME as CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
+import { STORE_NAME as MODULES_ANALYTICS } from '../../../analytics/datastore/constants';
+import { isValidAMPExperimentJSON } from '../../util';
+import Link from '../../../../components/link';
+import ErrorText from '../../../../components/error-text';
 
 const { useSelect, useDispatch } = Data;
