+++ assets/js/modules/optimize/components/common/AMPExperimentJSONField.test.js
@@ -20,10 +20,10 @@
  * Internal dependencies
  */
 import AMPExperimentJSONField from './AMPExperimentJSONField';
-import { render } from '../../../../../tests/js/test-utils';
-import { STORE_NAME } from '../datastore/constants';
-import { STORE_NAME as CORE_SITE } from '../../../googlesitekit/datastore/site/constants';
-import { STORE_NAME as MODULES_ANALYTICS } from '../../analytics/datastore/constants';
+import { render } from '../../../../../../tests/js/test-utils';
+import { STORE_NAME } from '../../datastore/constants';
+import { STORE_NAME as CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
+import { STORE_NAME as MODULES_ANALYTICS } from '../../../analytics/datastore/constants';
 
 describe( 'AMPExperimentJSONField', () => {
 	it( 'should render with all arguments passed', () => {
