+++ assets/js/modules/optimize/components/common/OptimizeIDField.test.js
@@ -20,9 +20,9 @@
  * Internal dependencies
  */
 import OptimizeIDField from './OptimizeIDField';
-import { render } from '../../../../../tests/js/test-utils';
-import { STORE_NAME } from '../datastore/constants';
-import { createTestRegistry } from '../../../../../tests/js/utils';
+import { render } from '../../../../../../tests/js/test-utils';
+import { STORE_NAME } from '../../datastore/constants';
+import { createTestRegistry } from '../../../../../../tests/js/utils';
 
 describe( 'OptimizeIDField', () => {
 	let registry;
