+++ assets/js/modules/optimize/components/common/index.js
