+++ assets/js/modules/optimize/components/settings/index.js
