+++ assets/js/modules/optimize/components/setup/SetupForm.js
@@ -31,10 +31,10 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import Button from '../../../components/button';
-import { STORE_NAME } from '../datastore/constants';
-import { isValidOptimizeID } from '../util';
-import ErrorText from '../../../components/error-text';
+import Button from '../../../../components/button';
+import { STORE_NAME } from '../../datastore/constants';
+import { isValidOptimizeID } from '../../util';
+import ErrorText from '../../../../components/error-text';
 import {
 	ErrorNotice,
 	AMPExperimentJSONField,
