+++ assets/js/modules/optimize/components/setup/index.js
