+++ assets/js/modules/optimize/index.js
@@ -25,8 +25,8 @@ import { addFilter } from '@wordpress/hooks';
  * Internal dependencies
  */
 import './datastore';
-import { SetupMain as OptimizeSetup } from './setup';
-import { SettingsMain as OptimizeSettings } from './settings';
+import { SetupMain as OptimizeSetup } from './components/setup';
+import { SettingsMain as OptimizeSettings } from './components/settings';
 import { fillFilterWithComponent } from '../../util';
 
 /**
