+++ stories/module-optimize-settings.stories.js
@@ -30,7 +30,7 @@ import { removeAllFilters, addFilter } from '@wordpress/hooks';
  * Internal dependencies
  */
 import SettingsModule from '../assets/js/components/settings/settings-module';
-import { SettingsMain as OptimizeSettings } from '../assets/js/modules/optimize/settings';
+import { SettingsMain as OptimizeSettings } from '../assets/js/modules/optimize/components/settings';
 import { fillFilterWithComponent } from '../assets/js/util';
 import { STORE_NAME as CORE_MODULE } from '../assets/js/googlesitekit/modules/datastore/constants';
 import { STORE_NAME as CORE_SITE } from '../assets/js/googlesitekit/datastore/site/constants';
